Begin to consolidate readline logic
